+++ documentation/changelog.rst
@@ -9,6 +9,7 @@ v0.32.0 | April XX, 2026
 
 New features
 -------------
+* Support saving state-of-charge schedules to sensors with ``"%"`` unit, using the ``soc-max`` flex-model field as the capacity for unit conversion [see `PR #2002 <https://www.github.com/FlexMeasures/flexmeasures/pull/2002>`_]
 
 Infrastructure / Support
 ----------------------